Towns and Fleeing!
When entering a fight the user now has an option to run away from the battle, on doing this they enter a town. This town can either be an enemy town or an allying town. If it is an enemy town there is a chance of you losing 10 health! If It's an allying town you could refill your armour or be healed!